+++ lib/edge/src/read_view/ops/count.rs
