+++ lib/edge/src/read_view/ops/facet.rs
