+++ lib/edge/src/read_view/ops/info.rs
