+++ lib/edge/src/read_view/ops/query.rs
