+++ lib/edge/src/read_view/ops/retrieve.rs
