+++ lib/edge/src/read_view/ops/search.rs
